requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via nicegui
anyascii==0.3.2
    # via wagtail
anyio==4.2.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asgiref==3.7.2
    # via django
asttokens==2.4.1
    # via stack-data
beautifulsoup4==4.11.2
    # via wagtail
bidict==0.22.1
    # via python-socketio
boto3==1.34.23
    # via
    #   django-ses
    #   django-storages
botocore==1.34.23
    # via
    #   boto3
    #   s3transfer
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   pip-tools
    #   uvicorn
coverage[toml]==7.4.0
    # via
    #   -r requirements.in
    #   pytest-cov
crispy-bootstrap3==2024.1
    # via -r requirements.in
cryptography==41.0.7
    # via django-ses
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via willow
django==5.0.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap3
    #   django-crispy-forms
    #   django-extensions
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-recaptcha
    #   django-ses
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   model-bakery
    #   wagtail
django-crispy-forms==2.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap3
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==23.5
    # via wagtail
django-image-uploader-widget==0.4.1
    # via -r requirements.in
django-modelcluster==6.2.1
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-recaptcha==3.0.0
    # via
    #   -r requirements.in
    #   wagtail-django-recaptcha
django-ses[events]==3.5.2
    # via -r requirements.in
django-storages[boto3]==1.14.2
    # via -r requirements.in
django-taggit==4.0.0
    # via wagtail
django-treebeard==4.7
    # via wagtail
djangorestframework==3.14.0
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
execnet==2.0.2
    # via pytest-xdist
executing==2.0.1
    # via stack-data
facebook-sdk==3.1.0
    # via -r requirements.in
factory-boy==3.3.0
    # via wagtail-factories
faker==22.4.0
    # via factory-boy
fastapi==0.109.0
    # via nicegui
filetype==1.2.0
    # via willow
freezegun==1.4.0
    # via pytest-freezegun
greenlet==3.0.3
    # via playwright
gunicorn==21.2.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
html5lib==1.1
    # via wagtail
htmx==0.0.0
    # via -r requirements.in
httpcore==1.0.2
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.26.0
    # via nicegui
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
ifaddr==0.2.0
    # via nicegui
iniconfig==2.0.0
    # via pytest
ipython==8.20.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via nicegui
jedi==0.19.1
    # via ipython
jinja2==3.1.3
    # via nicegui
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
l18n==2021.3
    # via wagtail
markdown2==2.4.10
    # via nicegui
markupsafe==2.1.4
    # via jinja2
matplotlib-inline==0.1.6
    # via ipython
model-bakery==1.17.0
    # via -r requirements.in
nicegui==1.4.11
    # via htmx
openpyxl==3.1.2
    # via wagtail
orjson==3.9.12
    # via nicegui
packaging==23.2
    # via
    #   build
    #   gunicorn
    #   pytest
parso==0.8.3
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.2.0
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.14.0
    # via willow
pip-tools==7.3.0
    # via -r requirements.in
playwright==1.41.0
    # via -r requirements.in
pluggy==1.3.0
    # via pytest
prompt-toolkit==3.0.43
    # via ipython
pscript==0.7.7
    # via vbuild
psycopg2==2.9.9
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pydantic==1.10.14
    # via
    #   fastapi
    #   htmx
pyee==11.0.1
    # via playwright
pygments==2.17.2
    # via
    #   ipython
    #   nicegui
pymemcache==4.0.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pytest==7.4.4
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-env
    #   pytest-freezegun
    #   pytest-xdist
pytest-cov==4.1.0
    # via -r requirements.in
pytest-django==4.7.0
    # via -r requirements.in
pytest-env==1.1.3
    # via -r requirements.in
pytest-freezegun==0.4.2
    # via -r requirements.in
pytest-xdist==3.5.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   faker
    #   freezegun
python-dotenv==1.0.0
    # via uvicorn
python-engineio==4.8.2
    # via python-socketio
python-multipart==0.0.6
    # via nicegui
python-socketio==5.11.0
    # via nicegui
pytz==2023.3.post1
    # via
    #   django-modelcluster
    #   django-ses
    #   djangorestframework
    #   l18n
pyyaml==6.0.1
    # via uvicorn
requests==2.31.0
    # via
    #   django-ses
    #   facebook-sdk
    #   wagtail
s3transfer==0.10.0
    # via boto3
simple-websocket==1.0.0
    # via python-engineio
six==1.16.0
    # via
    #   asttokens
    #   html5lib
    #   l18n
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
stack-data==0.6.3
    # via ipython
starlette==0.35.1
    # via fastapi
telepath==0.3.1
    # via wagtail
tqdm==4.66.1
    # via -r requirements.in
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.9.0
    # via
    #   fastapi
    #   nicegui
    #   pydantic
    #   pyee
urllib3==2.0.7
    # via
    #   botocore
    #   requests
uvicorn[standard]==0.26.0
    # via nicegui
uvloop==0.19.0
    # via uvicorn
vbuild==0.8.2
    # via nicegui
wagtail==5.2.2
    # via
    #   -r requirements.in
    #   wagtail-factories
    #   wagtail-json-widget
wagtail-django-recaptcha==1.0
    # via -r requirements.in
wagtail-factories==4.1.0
    # via -r requirements.in
wagtail-json-widget==0.0.7
    # via -r requirements.in
watchfiles==0.21.0
    # via
    #   nicegui
    #   uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
websockets==12.0
    # via uvicorn
wheel==0.42.0
    # via pip-tools
whitenoise==6.6.0
    # via -r requirements.in
willow[heif]==1.6.3
    # via wagtail
wsproto==1.2.0
    # via simple-websocket

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
